requirements.txt
# ── Core agent framework ───────────────────────────────────────
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain-core>=0.3.0
langchain-mcp-adapters>=0.1.0

# ── Async SQLite checkpointer ──────────────────────────────────
aiosqlite>=0.19.0

# ── MCP server runtime ─────────────────────────────────────────
mcp>=1.0.0
fastmcp>=0.1.0

# ── Infrastructure clients (used by MCP servers) ───────────────
kubernetes>=28.0.0
requests>=2.31.0

# ── CLI & config ───────────────────────────────────────────────
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
nest-asyncio>=1.6.0

# ── Testing ────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── LLM providers — uncomment the ones you need ───────────────
langchain-google-genai>=2.0.0      # Gemini  (GOOGLE_API_KEY)
langchain-openai>=0.2.0            # OpenAI + any OpenAI-compatible API
# langchain-anthropic>=0.3.0       # Claude  (ANTHROPIC_API_KEY)
# langchain-ollama>=0.2.0          # Ollama  (no key needed)
# langchain-mistralai>=0.2.0       # Mistral (MISTRAL_API_KEY)

# ── Docker SQLite backend (optional) ──────────────────────────
# Uncomment when SQLITE_URL is set to use the sqld container:
# libsql-experimental>=0.3.0
